Makefile: Speed up clone of git-core
When building Git ourselves, we remove any pre-existing Git clones and
clone the Git repository anew. It thus doesn't make a lot of sense to
clone a full repository, as it's going to be thrown away on rebuild
anyway. So let's convert the clone to be a shallow clone to speed up the
build.