ci: convert to use matrix builds
Our CI setup uses an implicit matrix to build, assemble and test our
code against different Go, Git and Ruby versions. In order to not repeat
ourselves, we thus used YAML anchors to merge in a common definition,
which already avoids quite some maintenance burden. But the end result
is still a tad hard to read as it's easy to miss the differences between
jobs.

Let's migrate to parallel matrix jobs instead. This new feature which
has been introducen in GitLab v13.3 allow us to do exactly what we want:
given a matrix of variables, run the job with each combination. This
allows us to get rid of some of the YAML anchors and explicitly states
what we're iterating over, improving maintainability.

Note that while the GIT_VERSION variable isn't yet used by anything, it
will get used as soon as we start using a custom-built Git version in
our CI pipelines. But as matrix builds require at least two variables,
they're added as part of this commit already.