Makefile: Stop using CI_JOB_NAME versions for generated reports
In order to make test results available to the CI, we generate JUnit
test reports. The generated files currently carry the CI_JOB_NAME
variable in their name, mostly to disambiguate them when downloading
multiple reports and extracting them. With migrating to matrix-based CI
jobs, the naming pattern has changed though to be e.g. "test 2/8". While
not being too helpful, it also carries a space and a slash character. As
a result, the report name is not referencing a file anymore, but instead
a directory hierarchy, causing the Makefile to fail.

Fix the issue by explicitly using the Go and Git versions as part of the
generated report name instead.