Failing unit test for hanging info ref cache
Introduces a failing test for when the info/ref cache encounters an
error while attempting to persist a response stream. This validates the
problem described in https://gitlab.com/gitlab-org/gitaly/-/issues/3087.

The failing test accomplishes overrides the info ref cache with a mock
implementation that always returns an error. The test asserts that a
normal response is received regardless of the error, and also asserts
that the mocked code was actually called.