Makefile: Run tests with self-built Git binary
Tests of Gitaly are currently always running against the system-provided
Git version. This is quite inflexible, as it doesn't allow for testing
against different versions of Git. As a workround, we provide a set of
different Docker images with different versions of Git, but this doesn't
really help developers who want to test against different versions on
their own machine.

This commit tries to improve the situation by making tests depend on our
new generic "git" rule and exporting the `GITALY_TESTING_GIT_BINARY`
environment variable. Like this, we'll always execute Git tests against
this version.