build libgit2 using git from path
`test` prepends an always failing git script to the path to fail tests
that use git from the path rather than the git binary provided explicitly
via the environment. `libgit2.a` target tries to use this always failing
script when building libgit2 as a dependency of `test`. This commit fixes
the problem by using the real git binary that was captured from the user's
path prior to overriding it.