hooks: Remove update feature flag
The update hooks was enabled by default, and as such is being used by
customers and GitLab.com in production. The feature is stable, and a
nice performance improvement. This change removes the feature gate, and
removes the option to disable the feature.

The biggest change are the cascading effects of removing a second
implemenation. This change does not yet remove the old hook as there
might be a race condition during deploy where the old hook is executed
while the file is removed. In such a case, the hook wouldn't be
executed. The old hook can be removed one release later.