hooks: Re-enable the reference-transaction hook
The reference-transaction hook included in Git v2.28.0 is currently
broken with interleaving hook calls, where the wrong hook would start to
be executed with parameters of the reference-transaction hook. As a
result, the commit introducing it got reverted.

At the point of writing, there is no release of Git including the fixed
version of that hook yet, and chances are good we'll have to wait
another two months until Git v2.29.0. To unblock ourselves, a workaround
has been added in c9d0ae5a (gitaly-hooks: Fixup reference-transaction
hook name based on arguments, 2020-08-31). Together with the testcase in
this commit's predecessor, we can now include the reference-transaction
hook again and be sure to not trigger the original issue anymore.

This reverts commit 19e2caa3a8a9fe390b568dd8d2b2a565be6094a7 and wires
up the reference-transaction hook again.