smarthttp: Push multiple branches in git-receive-pack tests
Until now, the PostReceivePack tests only ever pushed a single branch.
As recent breakage in the reference-transaction hook only gets triggered
in case there's interleaving hook calls, and there is no interleaving
calls with a single branch only, we weren't able to catch this breakage
with our tests.

Let's adjust tests to push multiple branches instead. This change
suffices to catch the regression if it were to happen again.