Makefile: Fix phony-ness of git target
While he "git" target is supposed to be a PHONY one, the preceding PHONY
declaration was missing the target's name. This is harmless in most
cases, but if one somehow ends up with a file named "git" in the
top-level directory, then the target could potentially become stale.

Fix this by adding the missing target to the PHONY declaration.