Improved SQL to get up to date storages for repository.
The SQL query to get up to date storages of the repository got
updated to one that shows better results. The benchmark test
added to check impact in future. The optimization is needed
as the resource consumption by the Postgresql instance increases
the more read operations praefect receives to serve.
It addresses the problem only partially as we should reduce
the load to database more with applying other optimizations.

Part of: https://gitlab.com/gitlab-org/gitaly/-/issues/3053