danger: Allow scoped merge request titles
In the Git ecosystem, commit titles are commonly scoped to a given
subsystem to make immediately clear which parts of the system are being
changed (see e.g. this very commit's title). Using the same pattern for
merge requests is thus a natural urge for many used to this, but our
Danger rules will complain about such commit titles as their first
character is not an uppercase on.e

Let's improve the rule to allow optional (lower-cased) scopes.