Makefile: Avoid Git2Go being linked against stale libgit2
Even though we're rebuilding libgit2.a on Makefile changes, this won't
update Git2Go. As a result, there may be a stale version of Git2Go in
the module cache that's linked against an old static archive of libgit2.

Fix this issue by always rebuilding Git2Go in case libgit2 gets rebuilt.