transactions: Enable majority-wins voting strategy by default
In order to get our feet wet with transactions, we initially enabled
transactions with a primary-wins voting strategy. This means that as
long as the primary casts a vote, the transaction is guaranteed to
be committed. This of course doesn't buy us any consistency guarantees,
but it did allow initial testing of the code without too much risk.

Given that primary-wins feature flag has been disabled in production for
some time now without any outfall at all, let's change its default value
to "disabled" to enable the majority-wins voting strategy by default.