gitaly: Refactor server setup to take a hook manager
The Gitaly server setup requires to pass a `hook.GitlabAPI` object which
is in turn used to instantiate the `hook.Manager`. This commit refactors
the setup to instead take a manager directly which allows us to inject
the manager from the top-level command similar like we do in Praefect.
One benefit is that we can thus move setup of metrics into the manager
itself and then register those in `cmd/gitaly`.

Note that this commit removes the voting delay metric altogether without
replacement. This is only done to not do too many things at the same
time, but it'll resurface in the commit immediately following this one.