hook: Reintroduce registration of voting delay metric
The previous commit has removed registration of the voting delay metric
in order to refactor how the hook manager is set up. With the
refactoring done now, we can now refactor the hook manager to set up the
metric itself and implement the `prometheus.Collector` interface such
that we can directly register the manager itself as a metric.