hook: Pass complete configuration to the hook manager
When instantiating a new hook manager, we currently pass only the hook
configuration to it as that's all we required up to now. But as we're
about to move the setup of Prometheus metrics into the hook manager,
we'll also require configuration of Prometheus buckets.

This commit thus refactors the hook manager to accept the complete
Gitaly configuration as parameter and adjusts any callers.