hook: Use manager's hook config to construct custom hooks executor
When creating custom hook executors, we pass in the repository path,
global hook path as well as the hook we want to execute. But in fact,
the global hook path is redundant as the hook manager already has it as
part of its `config.Hook` field.

Remove the redundant parameter to avoid any mishaps. This also removes
use of the global configuration in the update hook, where we didn't use
the manager's configuration by accident.