hook: Pass gitalypb.Repository into custom hooks executor
In order to execute custom hooks, we need to know which repository to
execute hooks in. We thus pass the repository path into the custom hook
executor's constructor, but as a result we've duplicated the logic to
retrieve that path at all callsites.

Refactor the constructor to accept a `gitalypb.Repository` instead of
its path to both deduplicate the logic and to have stronger typing.