hooks: Fix GitPushOptions for PostReceiveHook
Prior to this change the PostReceiveHook handler would attempt to use
the command environment to parse the Git options. This doesn't work
though, given the daemon would run in a different context than the
`gitaly-hook` gets executed in.

The Gitaly hook would parse the env correctly, pass it to the Gitaly
service through protobuf, the handler just didn't use it.