Parse response form to struct
The post receive hook tests used a map[string]interface{}. When an empty
interface is used, Golang has more trouble with type assertions, in this
case breaking `params["push_options"].([]string)` as the interface of a
slice is unlike other structs.

By parsing directly in a struct, the reflect package doesn't have to be
used, and we can depend on the type system some more.