Fix logging of replication events processing.
We should have an explicit markers of replication event
processing start and stop with the final state of the event.
It will help understand if processing of replication event
was done properly or some interruption happened as well
as identify events that take too long to be processed.
The starting log entry includes event because after successful processing
it will be removed and there is no way to check its target/source
nodes and repository. All replication log entries for event has
'correlation_id' so it is easy to get all related log entries by scanning
for it.

Part of: https://gitlab.com/gitlab-org/gitaly/-/issues/3047