Export GL_REPOSITORY and GL_PROJECT_PATH in `git archive` call
To support bundling Git LFS objects within a `git archive` call, we will
need a custom LFS smudge filter that will download the LFS data.  To do
this, we will need to make an API call request to check the the
requested LFS object ID actually belongs to the LFS object network, so
we need to export `GL_REPOSITORY` to the environment. `GL_PROJECT_PATH`
may be useful for debugging purposes.