git2go: Fix tests depending on local timezone
In order to test merges via gitaly-git2go, we're manually creating
temporary commits which we're merging with each other. These commits
currently have a `time.Unix(0, 0)` as both author and committer time.
But as we do not specify which timezone, this is dependent on the
timezone of the operating system and may thus fail on some systems, as
resulting commit hashes are not matching anymore.

Fix this by using an explicit time as well as an explicit timezone.