Protobuf generates different code locally and on CI
As it is hard to reason about what is the difference between
local run and run on CI in the Docker container the first
step to identify what is the actual difference between two.
If it is only a formatting then it is possible that locally
generated version was formatted and committed.
To see the difference a new artifact 'proto.diff' would be
available after failed execution of the 'no-changes' task.
It will show delta between locally generated and committed
files vs files generated on CI during verification.

Part of: https://gitlab.com/gitlab-org/gitaly/-/issues/2855