praefect: Add more tolerance in racy error threshold tests
As the error threshold checks function asynchronously, we're sleeping in
the tests which exercises that code. Naturally, CI still manages to be
slow enought to loose the race.

Improve the situation by increasing the amount of time waited from 10ms
to 50ms. To not penalize fast machines, we now sleep in a loop for 1ms
steps until we see the expected error.