hooks: Extract function to call hooks accepting stdin
As there are multiple hooks in Git which accept both arguments and input
from stdin, it makes sense to refactor calling such hooks into its own
function. So this commit pulls out most of the implementation of the
receive hooks and makes it generic to make it reusable for other hooks.