hooks: Send reference-transaction state to hook service
Besides its standard input, the reference-transaction hook also gets a
state as input via its first parameter. This state is one of "prepared",
"committed" or "aborted", telling the hook in which state the underlying
reference transaction currently is.

The information isn't currently available to the hook service. This
commit thus changes the ReferenceTransaction RPC to include another
field for the state, containing an enumeration of either of the above
states.