Pass correlation ID to hooks
Previously the `correlation_id` was not passed along the `X-Request-Id`
header to the GitLab API requests
(e.g. `/api/v4/internal/{pre_receive,post_receive}`). Now we extract the
`CORRELATION_ID` environment variable, store it in the context for the
gitaly-hooks binary, and pass the context along to the API requests.

This requires an update to gitlab-shell to support this:
https://gitlab.com/gitlab-org/gitlab-shell/-/merge_requests/413

Fixes https://gitlab.com/gitlab-org/gitaly/-/issues/2725