Allow null generations in repositories table
'repositories' table's records are used to keep record of repositories
that should exist and as repository specific locks for accessing
'storage_repositories' table.

The repository specific locking was implemented by incrementing the
generation column before modifying the entries in the 'storage_repositories'
table. Furthermore, the generation column is used as an easy way to
access the latest generation number. These uses prevent adding other
columns to the table, as inserting a record forces one to also set a
generation number from unrelated code. To make it viable to add other
repository specific columns to the table, this commit stops using the
'repositories.generation' as an easy way to get the current expected
generation and instead gets it by getting the maximum value from the
'storage_repositories' table. To make it possible to add other columns
without affecting the generation, NOT NULL constraint is dropped from
the column and queries have been updated to handle the case.

The column is still kept around and updated in the writes to keep
backwards compatibility with earlier versions. The column can be dropped
later once every repository is guaranteed to have a record in the table
by replacing the increment operation with a select for update.

Behavior should be semantically the same. Only difference is in
GetConsistentSecondaries which was only checking the secondaries match
the primary's generation rather than that they are on the latest generation.
This was changed as we can't conistency of the repository unless it
is on the very latest generation.