git2go: Implement command to list merge conflicts
In order to get a list of all conflicting files when merging two
commits, this commit implements a new command `gitaly-git2go conflicts`.
As input, it gets the repository as well as the two commits which ought
to be merged with each other. It'll then return the serialized list of
conflicts via its standard output, containing for each conflict the
three paths which are impacted as well as the conflicting hunks.