docs: Explain how to interact with Git in Gitaly
When interacting with Git repositories the Gitaly team contains a lot of
knowledge, but it's not written down. This makes it harder for
maintainers to spread their knowlegde, and would ease reviews if
contributors can learn what to do and not to do.

This would also allow maintainers to link to the docs during reviews
without explaining the full story, and speed them up.