git2go: Make conflict entry's modes available
When listing merge conflicts via Git2Go, we currently only pass along
information about conflict paths and contents. In order to implement
ListConflictFiles, we also require the mode of each of the sides.

Add the mode to the conflicts response to cater to this need. In order
to avoid code duplication, this commit also does some refactoring and
sets up a separate structure for conflict entries.