transactions: Correctly handle cancellation of empty transactions
If cancelling a transaction which doesn't have any subtransactons, we
currently loose track of the fact that the transaction has been
cancelled as we only use subtransaction information to determine the
transaction's state. It's thus fine to vote on an cancelled, but empty
transaction, which is simply wrong.

Fix this by explicitly keeping track of the cancellation state, refusing
to create any new subtransactions if the transaction is in such a state.