git2go: Serialize results directly into os.Stdout
When spawning Git2Go commands, communication happens via serialized
messages between parent and its child process. This is currently done by
serializing into a string, only to print that string to standard output
immediately afterwards. This pattern is quite inefficient, as we first
need to serialize the whole string into memory.

This commit implements a new function to serialize directly into a
reader in order to avoid this overhead and adjusts the merge command to
use it.