git2go: Store conflict contents as bytes
When conflicts occur, it's not necessarily the case that the conflict's
contents are valid UTF-8, as conflicting files may have different
encodings. But as we're serializing and deserializing those contents as
strings, any such encoding issues may get stripped away and cause us to
end up with conflict contents which don't really represent the actual
conflict.

Fix this by using a byte array for conflict contents instead of a
string.