Makefile: Fix check-formatting not calling goimports
The check-formatting target is supposed to verify whether files are
formatted correctly. In the best case, the target would complain
whenever the format target would've rewritten files. Turns out this is
not the case as it calls the `gitalyfmt` command, but not `goimports`
like the format target does. As a result, `make format` may fix
formatting even though `make check-formatting` would pretend everything
is alright.

Fix the issue by also calling `goimports -l` when checking the format.