transactions: Allow graceful stop of transactions
It is currently only possible to stop transactions by cancelling them,
which will cause all ongoing and future votes on this transaction to
raise errors. But in some cases we need to stop transactions gracefully
without raising errors, which is nothing that can be achieved right now.

This commit implements the first piece of the puzzle to allow for this
use case, which is stopping of transactions. Semantics are mostly the
same as for transaction cancellation, but with proper error handling and
state tracking such that multiple calls to stop will not panic.