demo: Verify Gitaly install with gitaly-hooks
The terraform install uses `bin/check` from the vendorred `gitlab-shell`
code to check the hooks are configured correctly. By changing this to
`gitaly-hooks` we can prepare for a later situation where `gitlab-shell`
is removed and as such not available.

Currently the `gitaly-hooks check` command executes `gitlab-shell` under
the hood, so the results are identical.