PostgreSQL notifications listener
Praefect strongly depends on the state of tha PostgreSQL database.
Usually we execute a query to get the latest infornation about the
state. But it doesn't scale well in case query is heavy or there are
too many round trips to the database. To address it the data could be
cached locally in the memory, so we can reduce amount of trips to the
database. The downside of this approach is that we should invalidate
local cache when data in database changes.

To address this we decided to allow Praefect to have a direct connection
to the PostgeSQL database and benefit from LISTEN/NOTIFY functionality of
it. Implemented listener triggers passed in callback function with payload
it receives from the database. In order to prevent stale state if connection
was lost and it can't be re-establish for 3 attempts in period of 1 min the
listener will return with error.

Part of: https://gitlab.com/gitlab-org/gitaly/-/issues/3053