git: Fix parsing of dashed -rc versions
Typically, Git release candidates use a dashed "-rc0" suffix to indicate
the prerelease status. We currently fail to parse this correctly as we
expect the rc suffix to be separated by a dot instead, e.g. "2.29.0.rc0"
instead of the actual "2.29.0-rc0".

Fix the issue by correctly parsing the dashed form.