danger: Allow numbers in MR's title scope
An MRs title is allowed to have a scope like "praefect:" as prefix,
which helps narrowing down which area of the code is being changed. The
rule is quite strict though and doesn't allow for numbers in the scope,
but given our recent trips into Git2Go land this case does come up quite
often now.

Make the rule more lax and allow numbers as part of the scope.