hook: Pull out logic to conditionally execute function on transaction
Voting on a transaction always depends on whether the current context
actually knows about transactions or not. As transactions aren't
and never will be mandatory, absence of a transaction isn't treated as a
failure case. There is thus common logic for acting on a transaction
which handles these cases.

This commit refactors the hook transaction logic to pull out this common
code as a preparatory step for stopping transactions.